adding scoring outline
